Renamed connection_provider.py file to client.py
