+++ pulumi_snowflake/database/database.py
@@ -4,7 +4,7 @@ from pulumi import Output, Input, ResourceOptions
 from pulumi.dynamic import Resource
 
 from ..provider import Provider
-from ..connection_provider import Client
+from ..client import Client
 from .database_provider import DatabaseProvider
